rpc: truncate num_blocks in create_pmem_pool
Python2 implicitly truncated the division of pool_size/block_size. If
there is any remainder in pyhton3, the value stored in num_blocks is
kept as a float. The cast to integer truncates this division resulting
in the same behavior between python2 and python3

+++ scripts/rpc/pmem.py
@@ -1,5 +1,6 @@
 def create_pmem_pool(client, args):
-    num_blocks = (args.total_size * 1024 * 1024) / args.block_size
+    # truncate to the nearest block.
+    num_blocks = int((args.total_size * 1024 * 1024) / args.block_size)
     params = {'pmem_file': args.pmem_file,
               'num_blocks': num_blocks,
               'block_size': args.block_size}
